Fix button disabled bug #178
The forEach() method does not handle asynchronous functions properly. It is not designed to work with async/await syntax or handle promises returned by asynchronous operations. This caused the checkAllSelectsHaveValue call in the fetchOptions method to be called before all results from the async have been returned resulting in the button to be left as disabled. Updated the fetchOptions method to with for all the search results to come back before calling the checkAllSelectsHaveValue method.